+++ app/app/main.py
@@ -19,11 +19,11 @@
 # ECOMP is a trademark and service mark of AT&T Intellectual Property.
 
 import connexion
-from config_binding_service.logging import create_logger
+from config_binding_service.logging import create_loggers
 
 # Entrypoint When in uwsgi
 # This create logger call used to be in the main block, but when moving to NGINX+uwsgi, this had to change. See https://hub.docker.com/r/tiangolo/uwsgi-nginx-flask/
-create_logger()
+create_loggers()
 app = connexion.App(__name__, specification_dir='.')
 app.add_api('swagger.yaml', arguments={'title': 'Config Binding Service'})